adding the current website undr UAT in repository

+++ controllers/ContactsController.php
@@ -12,13 +12,12 @@ use yii\filters\VerbFilter;
 /**
  * ContactsController implements the CRUD actions for Contacts model.
  */
-class ContactsController extends Controller
-{
+class ContactsController extends Controller {
+
     /**
      * @inheritdoc
      */
-    public function behaviors()
-    {
+    public function behaviors() {
         return [
             'verbs' => [
                 'class' => VerbFilter::className(),
@@ -28,19 +27,23 @@ class ContactsController extends Controller
             ],
         ];
     }
+    
+     public function init() {
+        $this->layout = 'backend/main';
+        parent::init();
+    }
 
     /**
      * Lists all Contacts models.
      * @return mixed
      */
-    public function actionIndex()
-    {
+    public function actionIndex() {
         $searchModel = new ContactsSearch();
         $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
         return $this->render('index', [
-            'searchModel' => $searchModel,
-            'dataProvider' => $dataProvider,
+                    'searchModel' => $searchModel,
+                    'dataProvider' => $dataProvider,
         ]);
     }
 
@@ -49,10 +52,9 @@ class ContactsController extends Controller
      * @param integer $id
      * @return mixed
      */
-    public function actionView($id)
-    {
+    public function actionView($id) {
         return $this->render('view', [
-            'model' => $this->findModel($id),
+                    'model' => $this->findModel($id),
         ]);
     }
 
@@ -61,17 +63,25 @@ class ContactsController extends Controller
      * If creation is successful, the browser will be redirected to the 'view' page.
      * @return mixed
      */
-    public function actionCreate()
-    {
+    public function actionCreate() {
         $model = new Contacts();
-
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->Id]);
-        } else {
-            return $this->render('create', [
-                'model' => $model,
-            ]);
+        if (Yii::$app->session->get('UNIT_ID')) {
+            $model->UnitID = Yii::$app->session->get('UNIT_ID');
         }
+
+        if ($model->load(Yii::$app->request->post())) {
+            if (Yii::$app->request->post('save') == 'save') {
+                $model->Status = Contacts::STATUS_SAVED;
+            } elseif (Yii::$app->request->post('publish') == 'publish') {
+                $model->Status = Contacts::STATUS_PUBLISHED;
+            }
+            if ($model->save()) {
+                return $this->redirect(['view', 'id' => $model->Id]);
+            }
+        }
+        return $this->render('create', [
+                    'model' => $model,
+        ]);
     }
 
     /**
@@ -80,17 +90,27 @@ class ContactsController extends Controller
      * @param integer $id
      * @return mixed
      */
-    public function actionUpdate($id)
-    {
+    public function actionUpdate($id) {
         $model = $this->findModel($id);
-
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->Id]);
-        } else {
-            return $this->render('update', [
-                'model' => $model,
-            ]);
+        if ($model -> Status == Contacts::STATUS_PUBLISHED) {
+            $this->redirect(array('index'));
         }
+        if (Yii::$app->session->get('UNIT_ID')) {
+            $model->UnitID = Yii::$app->session->get('UNIT_ID');
+        }
+        if ($model->load(Yii::$app->request->post())) {
+            if (Yii::$app->request->post('save') == 'save') {
+                $model->Status = Contacts::STATUS_SAVED;
+            } elseif (Yii::$app->request->post('publish') == 'publish') {
+                $model->Status = Contacts::STATUS_PUBLISHED;
+            }
+            if ($model->save()) {
+                return $this->redirect(['view', 'id' => $model->Id]);
+            }
+        }
+        return $this->render('update', [
+                    'model' => $model,
+        ]);
     }
 
     /**
@@ -99,10 +119,11 @@ class ContactsController extends Controller
      * @param integer $id
      * @return mixed
      */
-    public function actionDelete($id)
-    {
-        $this->findModel($id)->delete();
-
+    public function actionDelete($id) {
+        $model = $this->findModel($id);
+        if ($model-> Status == Contacts::STATUS_SAVED OR $model->Status == Contacts::STATUS_UNPUBLISHED) {
+            $this->findModel($id)->delete();
+        }
         return $this->redirect(['index']);
     }
 
@@ -113,12 +134,30 @@ class ContactsController extends Controller
      * @return Contacts the loaded model
      * @throws NotFoundHttpException if the model cannot be found
      */
-    protected function findModel($id)
-    {
+    protected function findModel($id) {
         if (($model = Contacts::findOne($id)) !== null) {
             return $model;
         } else {
             throw new NotFoundHttpException('The requested page does not exist.');
         }
     }
+
+    function actionPublish($id) {
+        $model = Contacts::findOne($id);
+        if ($model->Status == Contacts::STATUS_SAVED OR $model->Status == Contacts::STATUS_UNPUBLISHED) {
+            $model->Status = Contacts::STATUS_PUBLISHED;
+            $model->save();
+        }
+        return $this->redirect(['view', 'id' => $model->Id]);
+    }
+
+    function actionUnpublish($id) {
+        $model = Contacts::findOne($id);
+        if ($model->Status == Contacts::STATUS_PUBLISHED) {
+            $model->Status = Contacts::STATUS_UNPUBLISHED;
+            $model->save();
+        }
+        return $this->redirect(['view', 'id' => $model->Id]);
+    }
+
 }
