+++ controllers/ResearchProjectsController.php
@@ -12,13 +12,12 @@ use yii\filters\VerbFilter;
 /**
  * ResearchProjectsController implements the CRUD actions for ResearchProjects model.
  */
-class ResearchProjectsController extends Controller
-{
+class ResearchProjectsController extends Controller {
+
     /**
      * @inheritdoc
      */
-    public function behaviors()
-    {
+    public function behaviors() {
         return [
             'verbs' => [
                 'class' => VerbFilter::className(),
@@ -29,18 +28,25 @@ class ResearchProjectsController extends Controller
         ];
     }
 
+    
+     public function init() {
+        $this->layout = 'backend/main';
+        parent::init();
+    }
     /**
      * Lists all ResearchProjects models.
      * @return mixed
      */
-    public function actionIndex()
-    {
+    public function actionIndex() {
         $searchModel = new ResearchProjectsSearch();
+        if (\Yii::$app->session->has('UNIT_ID')) {
+            $searchModel->UnitID = \Yii::$app->session->has('UNIT_ID');
+        }
         $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
         return $this->render('index', [
-            'searchModel' => $searchModel,
-            'dataProvider' => $dataProvider,
+                    'searchModel' => $searchModel,
+                    'dataProvider' => $dataProvider,
         ]);
     }
 
@@ -49,10 +55,9 @@ class ResearchProjectsController extends Controller
      * @param integer $id
      * @return mixed
      */
-    public function actionView($id)
-    {
+    public function actionView($id) {
         return $this->render('view', [
-            'model' => $this->findModel($id),
+                    'model' => $this->findModel($id),
         ]);
     }
 
@@ -61,17 +66,31 @@ class ResearchProjectsController extends Controller
      * If creation is successful, the browser will be redirected to the 'view' page.
      * @return mixed
      */
-    public function actionCreate()
-    {
+    public function actionCreate() {
         $model = new ResearchProjects();
-
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->Id]);
-        } else {
-            return $this->render('create', [
-                'model' => $model,
-            ]);
+        if ($model->load(Yii::$app->request->post())) {
+            var_dump($model->attributes);
+            if ($model->ProjectNameEn) {
+                $model->ProjectLinkUrl = \app\components\Utilities::createUrlString($model->ProjectNameEn);
+            }
+            if (\Yii::$app->session->has('UNIT_ID')) {
+                $model->UnitID = \Yii::$app->session->has('UNIT_ID');
+            }
+            if (Yii::$app->request->post('save') == 'save') {
+                $model->Status = ResearchProjects::PROJECT_STATUS_SAVED;
+            } elseif (Yii::$app->request->post('publish') == 'publish') {
+                $model->Status = ResearchProjects::PROJECT_STATUS_PUBLISHED;
+            }
+            if ($model->ProjectNameEn) {
+                $model->PageLinkUrl = \app\components\Utilities::createUrlString($model->ProjectNameEn);
+            }
+            if ($model->save()) {
+                return $this->redirect(['view', 'id' => $model->Id]);
+            }
         }
+        return $this->render('create', [
+                    'model' => $model,
+        ]);
     }
 
     /**
@@ -80,17 +99,27 @@ class ResearchProjectsController extends Controller
      * @param integer $id
      * @return mixed
      */
-    public function actionUpdate($id)
-    {
+    public function actionUpdate($id) {
         $model = $this->findModel($id);
-
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->Id]);
-        } else {
-            return $this->render('update', [
-                'model' => $model,
-            ]);
+        if ($model->Status == ResearchProjects::PROJECT_STATUS_PUBLISHED) {
+            return $this->redirect(['index']);
         }
+        if ($model->load(Yii::$app->request->post())) {
+            if (Yii::$app->request->post('save') == 'save') {
+                $model->Status = ResearchProjects::PROJECT_STATUS_SAVED;
+            } elseif (Yii::$app->request->post('publish') == 'publish') {
+                $model->Status = ResearchProjects::PROJECT_STATUS_PUBLISHED;
+            }
+            if ($model->ProjectNameEn) {
+                $model->PageLinkUrl = \app\components\Utilities::createUrlString($model->ProjectNameEn);
+            }
+            if ($model->save()) {
+                return $this->redirect(['view', 'id' => $model->Id]);
+            }
+        }
+        return $this->render('update', [
+                    'model' => $model,
+        ]);
     }
 
     /**
@@ -99,10 +128,11 @@ class ResearchProjectsController extends Controller
      * @param integer $id
      * @return mixed
      */
-    public function actionDelete($id)
-    {
-        $this->findModel($id)->delete();
-
+    public function actionDelete($id) {
+        $model = $this->findModel($id);
+        if ($model->Status != ResearchProjects::PROJECT_STATUS_PUBLISHED) {
+            $this->findModel($id)->delete();
+        }
         return $this->redirect(['index']);
     }
 
@@ -113,12 +143,30 @@ class ResearchProjectsController extends Controller
      * @return ResearchProjects the loaded model
      * @throws NotFoundHttpException if the model cannot be found
      */
-    protected function findModel($id)
-    {
+    protected function findModel($id) {
         if (($model = ResearchProjects::findOne($id)) !== null) {
             return $model;
         } else {
             throw new NotFoundHttpException('The requested page does not exist.');
         }
     }
+
+    function actionPublish($id) {
+        $model = ResearchProjects::findOne($id);
+        if ($model->Status == ResearchProjects::PROJECT_STATUS_SAVED OR $model->Status == ResearchProjects::PROJECT_STATUS_UNPUBLISHED) {
+            $model->Status = ResearchProjects::PROJECT_STATUS_PUBLISHED;
+            $model->save();
+        }
+        return $this->redirect(['view', 'id' => $model->Id]);
+    }
+
+    function actionUnpublish($id) {
+        $model = ResearchProjects::findOne($id);
+        if ($model->Status == ResearchProjects::PROJECT_STATUS_PUBLISHED) {
+            $model->Status = ResearchProjects::PROJECT_STATUS_UNPUBLISHED;
+            $model->save();
+        }
+        return $this->redirect(['view', 'id' => $model->Id]);
+    }
+
 }
